Updated parallel client tunneling implementation, added test. Renamed tunnel modules and tests.
