+++ pssh/clients/base/single.py
@@ -59,6 +59,8 @@ class BaseSSHClient(object):
                  allow_agent=True, timeout=None,
                  proxy_host=None,
                  proxy_port=None,
+                 proxy_user=None,
+                 proxy_password=None,
                  _auth_thread_pool=True,
                  identity_auth=True):
         self._auth_thread_pool = _auth_thread_pool
@@ -78,6 +80,8 @@ class BaseSSHClient(object):
         self.session = None
         self._host = proxy_host if proxy_host else host
         self._port = proxy_port if proxy_port else port
+        self._user = proxy_user if proxy_user else user
+        self._password = proxy_password if proxy_password else password
         self.pkey = _validate_pkey_path(pkey, self.host)
         self.identity_auth = identity_auth
         self._keepalive_greenlet = None
